Move models from nanny_models to models file.
