+++ application/models/__init__.py
@@ -1,14 +1,14 @@
-from .nanny_models.declaration import *
-from .nanny_models.nanny_application import *
-from .nanny_models.childcare_training import *
-from .nanny_models.childcare_address import *
-from .nanny_models.base import *
-from .nanny_models.first_aid import *
-from .nanny_models.applicant_personal_details import *
-from .nanny_models.applicant_home_address import *
-from .nanny_models.insurance_cover import *
-from .nanny_models.dbs_check import *
-from .nanny_models.application_reference import *
-from .nanny_models.payment import *
+from application.models.declaration import *
+from application.models.nanny_application import *
+from application.models.childcare_training import *
+from application.models.childcare_address import *
+from application.models.base import *
+from application.models.first_aid import *
+from application.models.applicant_personal_details import *
+from application.models.applicant_home_address import *
+from application.models.insurance_cover import *
+from application.models.dbs_check import *
+from application.models.application_reference import *
+from application.models.payment import *
 
 name = "nanny_models"
