+++ application/models/childcare_training.py
@@ -1,7 +1,6 @@
 from uuid import uuid4
 
 from rest_framework import serializers
-from .base import ApiCalls
 from django.db import models
 
 from .nanny_application import NannyApplication
@@ -12,7 +11,6 @@ class ChildcareTraining(models.Model):
     Model for Childcare Training table.
     """
     objects = models.Manager()
-    api = ApiCalls("childcare-training", 'childcare_training_id')
 
     childcare_training_id = models.UUIDField(primary_key=True, default=uuid4)
     application_id = models.ForeignKey(NannyApplication, on_delete=models.CASCADE, db_column='application_id')
